aadl-translator/src/main/java/edu/ksu/cis/projects/mdcf/aadltranslator/model/ProcessModel.java
package edu.ksu.cis.projects.mdcf.aadltranslator.model;

import java.util.HashMap;

import edu.ksu.cis.projects.mdcf.aadltranslator.exception.DuplicateElementException;
import edu.ksu.cis.projects.mdcf.aadltranslator.model.ModelUtil.ComponentType;
import edu.ksu.cis.projects.mdcf.aadltranslator.model.ModelUtil.ProcessType;

public class ProcessModel extends DevOrProcComponentModel{
	
	// variable name -> type
	private HashMap<String, String> globals;
	
	// method name -> method model
	private HashMap<String, MethodModel> methods;

	public ProcessModel() {
		super();
		methods = new HashMap<>();
		globals = new HashMap<>();
	}

	public HashMap<String, String> getGlobals() {
		return globals;
	}
	
	/**
	 * This will return the type of a global variable
	 * @param name The name of the global variable
	 * @return The type of the global variable
	 */
	public String getGlobalType(String name){
		return globals.get(name);
	}

	public void setDisplay(boolean display) {
		if(display){
			this.processType = ProcessType.DISPLAY;
			this.componentType = ComponentType.ACTUATOR;
		} else {
			this.processType = ProcessType.LOGIC;
			this.componentType = ComponentType.CONTROLLER;
		}
	}

	public TaskModel getLastThread(){
		return (TaskModel) children.get(children.size() - 1);
	}
	
	public void addGlobal(String name, String type) {
		globals.put(name, type);
	}

	public HashMap<String, MethodModel> getMethods() {
		return methods;
	}

	public void addMethod(String methodName, MethodModel method) {
		methods.put(methodName, method);
	}
	
	public void addParameterToMethod(String methodName, String parameterName, String parameterType) throws DuplicateElementException{
		if(!methods.containsKey(methodName))
			methods.put(methodName, new MethodModel(methodName));
		methods.get(methodName).addParameter(parameterName, parameterType);
	}

	public void addReturnToMethod(String methodName, String returnType) {
		if(!methods.containsKey(methodName))
			methods.put(methodName, new MethodModel(methodName));
		methods.get(methodName).setRetType(returnType);
	}
}
